pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "py-zippy"
version = "1.0.0"
description = "Archive Utility Toolkit supporting multi-format workflows, password recovery, and repair helpers."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
authors = [
    { name = "John Hauger Mitander", email = "john@on1.no" },
]
keywords = ["archives", "zip", "tar", "cli", "toolkit"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Archiving",
    "Topic :: System :: Archiving :: Compression",
    "Topic :: Utilities",
]
dependencies = [
    "colorama>=0.4.6",
    "python-dotenv>=1.0",
    "pyzipper>=0.3.6",
    "patool>=1.12",
]

[project.urls]
Homepage = "https://github.com/John0n1/ZIPPY"
Documentation = "https://github.com/John0n1/ZIPPY#readme"
Issues = "https://github.com/John0n1/ZIPPY/issues"

[project.scripts]
zippy = "zippy.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["zippy", "zippy.*"]
